Refactor import logic to standalone manager
