+++ src/main/java/seedu/contax/logic/commands/ImportCsvCommand.java
@@ -8,25 +8,12 @@ import static seedu.contax.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.contax.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.contax.logic.parser.CliSyntax.PREFIX_TAG;
 
-import java.io.BufferedReader;
-import java.io.FileReader;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Set;
 
 import seedu.contax.logic.commands.exceptions.CommandException;
-import seedu.contax.logic.parser.ParserUtil;
-import seedu.contax.logic.parser.exceptions.ParseException;
 import seedu.contax.model.IndexedCsvFile;
 import seedu.contax.model.Model;
-import seedu.contax.model.person.Address;
-import seedu.contax.model.person.Email;
-import seedu.contax.model.person.Name;
-import seedu.contax.model.person.Person;
-import seedu.contax.model.person.Phone;
-import seedu.contax.model.person.exceptions.DuplicatePersonException;
-import seedu.contax.model.tag.Tag;
+import seedu.contax.storage.CsvManager;
 
 public class ImportCsvCommand extends Command {
     public static final String COMMAND_WORD = "importcsv";
@@ -58,60 +45,10 @@ public class ImportCsvCommand extends Command {
     public CommandResult execute(Model model) throws CommandException {
         //process file
         try {
-            String line = "";
-            ArrayList<Integer> skippedLines = new ArrayList<>();
-            int lineCounter = 0;
-            BufferedReader importedCsv = new BufferedReader(new FileReader(toImport.getFilePath()));
-
-            //skip first line by default as it contains headers
-            importedCsv.readLine();
-
-            while ((line = importedCsv.readLine()) != null) {
-                lineCounter++;
-                String[] importedPerson = line.split(",");
-                try {
-                    Person toAddPerson = personParser(importedPerson);
-                    model.addPerson(toAddPerson);
-                    Set<Tag> tags = toAddPerson.getTags();
-                    for (Tag tag : tags) {
-                        if (!model.hasTag(tag)) {
-                            model.addTag(tag);
-                        }
-                    }
-                } catch (ParseException | DuplicatePersonException e) {
-                    skippedLines.add(lineCounter);
-                    continue;
-                }
-            }
-            return outputStringBuilder(skippedLines);
+            CsvManager manager = new CsvManager(model);
+            return manager.importCsv(toImport);
         } catch (IOException e) {
             throw new CommandException(String.format(MESSAGE_NO_FILE_FOUND, toImport.getFilePath()));
-        }
-    }
-    private Person personParser(String[] importedPerson) throws ParseException {
-        Name toAddName = ParserUtil.parseName(importedPerson[toImport.getNamePositionIndex()]);
-        Phone toAddPhone = ParserUtil.parsePhone(importedPerson[toImport.getPhonePositionIndex()]);
-        Email toAddEmail = ParserUtil.parseEmail(importedPerson[toImport.getEmailPositionIndex()]);
-        Address toAddAddress = ParserUtil.parseAddress(importedPerson[toImport.getAddressPositionIndex()]);
-        String[] tags = importedPerson[toImport.getTagPositionIndex()].split(";");
-        Set<Tag> toAddTag = ParserUtil.parseTags(Arrays.asList(tags));
-
-        return new Person(toAddName, toAddPhone, toAddEmail, toAddAddress, toAddTag);
-    }
-
-    private CommandResult outputStringBuilder(ArrayList<Integer> skippedLines) {
-        if (skippedLines.size() > 0) {
-            String skippedLinesString = "";
-            for (int i = 0; i < skippedLines.size(); i++) {
-                skippedLinesString += skippedLines.get(i);
-                if (i != skippedLines.size() - 1) {
-                    skippedLinesString += ", ";
-                }
-            }
-            return new CommandResult(String.format("%s\n%s", MESSAGE_SUCCESS,
-                    String.format(MESSAGE_SKIPPED_LINES, skippedLinesString)));
-        } else {
-            return new CommandResult(MESSAGE_SUCCESS);
         }
     }
